Ensured only move semantics are used in Trace and Result
