+++ src/engine/test/source/hlp2/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_executable(parsec_test
-    parsec_test.cpp
+    parsecTrace_test.cpp
+    parsecResult_test.cpp
+    parsecCombinators_test.cpp
 )
 target_link_libraries(parsec_test hlp::parsec gtest_main)
 gtest_discover_tests(parsec_test)
