+++ src/engine/test/source/hlp2/parsecCombinators_test.cpp
@@ -1,8 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <stdexcept>
-
-#include <hlp/parsec.hpp>
+#include <hlp/combinators.hpp>
 
 using resT = int;
 static const resT valTest = 1;
@@ -38,254 +36,6 @@ parsec::Parser<resT> getAnyParser()
     };
 }
 
-/****************************************************************************************/
-// Trace type tests
-/****************************************************************************************/
-TEST(ParsecTraceTest, BuildsDefault)
-{
-    ASSERT_NO_THROW(parsec::Trace {});
-}
-
-TEST(ParsecTraceTest, Builds)
-{
-    ASSERT_NO_THROW(parsec::Trace(true, 0, {}, {}));
-}
-
-TEST(ParsecTraceTest, BuildsCopy)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_NO_THROW(parsec::Trace {trace});
-    ASSERT_EQ(trace, parsec::Trace {trace});
-}
-
-TEST(ParsecTraceTest, BuildsMove)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    parsec::Trace expected {trace};
-    ASSERT_EQ(expected, parsec::Trace {std::move(trace)});
-}
-
-TEST(ParsecTraceTest, AssignationCopy)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    parsec::Trace trace2;
-    ASSERT_NO_THROW(trace2 = trace);
-    ASSERT_EQ(trace, trace2);
-}
-
-TEST(ParsecTraceTest, AssignationMove)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    parsec::Trace expected {trace};
-    ASSERT_EQ(expected, parsec::Trace {std::move(trace)});
-}
-
-TEST(ParsecTraceTest, Equality)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    parsec::Trace trace2 {true, 0, {}, {}};
-    ASSERT_EQ(trace, trace2);
-}
-
-TEST(ParsecTraceTest, Inequality)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    parsec::Trace trace2 {false, 0, {}, {}};
-    ASSERT_NE(trace, trace2);
-}
-
-TEST(ParsecTraceTest, Success)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_TRUE(trace.success());
-
-    trace = parsec::Trace {false, 0, {}, {}};
-    ASSERT_FALSE(trace.success());
-}
-
-TEST(ParsecTraceTest, Index)
-{
-    parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_EQ(0, trace.index());
-
-    trace = parsec::Trace {true, 1, {}, {}};
-    ASSERT_EQ(1, trace.index());
-}
-
-TEST(ParsecTraceTest, MessageRef)
-{
-    const parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_FALSE(trace.message());
-
-    const parsec::Trace trace1 = parsec::Trace {true, 0, "message", {}};
-    ASSERT_TRUE(trace1.message());
-    ASSERT_EQ("message", trace1.message().value());
-}
-
-TEST(ParsecTraceTest, MessageMove)
-{
-    auto getMessage = [](std::optional<std::string>&& message)
-    {
-        return message;
-    };
-
-    parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_FALSE(getMessage(trace.message()));
-
-    trace = parsec::Trace {true, 0, "message", {}};
-    std::optional<std::string> got;
-    ASSERT_TRUE(got = getMessage(trace.message()));
-    ASSERT_EQ("message", got.value());
-}
-
-TEST(ParsecTraceTest, InnerTracesRef)
-{
-    const parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_FALSE(trace.innerTraces());
-
-    const parsec::Trace trace1 = parsec::Trace {true, 0, {}, {{{true, 0, {}, {}}}}};
-    ASSERT_TRUE(trace1.innerTraces());
-    ASSERT_EQ(1, trace1.innerTraces().value().size());
-}
-
-TEST(ParsecTraceTest, InnerTracesMove)
-{
-    auto getInnerTraces = [](parsec::Trace::nestedTracesT&& innerTraces)
-    {
-        return innerTraces;
-    };
-
-    parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_FALSE(getInnerTraces(trace.innerTraces()));
-
-    trace = parsec::Trace {true, 0, {}, {{{true, 0, {}, {}}}}};
-    parsec::Trace::nestedTracesT got;
-    ASSERT_TRUE(got = getInnerTraces(trace.innerTraces()));
-    ASSERT_EQ(1, got.value().size());
-}
-
-/****************************************************************************************/
-// Result type tests
-/****************************************************************************************/
-TEST(ParsecResultTest, BuildsDefault)
-{
-    ASSERT_NO_THROW(parsec::Result<resT> {});
-}
-
-TEST(ParsecResultTest, BuildsParameters)
-{
-    ASSERT_NO_THROW(parsec::Result<resT>(valTest, {true, 0, {}, {}}));
-}
-
-TEST(ParsecResultTest, BuildsCopy)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    ASSERT_NO_THROW(parsec::Result<resT> {result});
-    ASSERT_EQ(result, parsec::Result<resT> {result});
-}
-
-TEST(ParsecResultTest, BuildsMove)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    parsec::Result<resT> expected {result};
-    ASSERT_EQ(expected, parsec::Result<resT> {std::move(result)});
-}
-
-TEST(ParsecResultTest, AssignationCopy)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    parsec::Result<resT> result2;
-    ASSERT_NO_THROW(result2 = result);
-    ASSERT_EQ(result, result2);
-}
-
-TEST(ParsecResultTest, AssignationMove)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    parsec::Result<resT> expected {result};
-    ASSERT_EQ(expected, parsec::Result<resT> {std::move(result)});
-}
-
-TEST(ParsecResultTest, Equality)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    parsec::Result<resT> result2 {valTest, {true, 0, {}, {}}};
-    ASSERT_EQ(result, result2);
-}
-
-TEST(ParsecResultTest, Inequality)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    parsec::Result<resT> result2 {valTest, {false, 0, {}, {}}};
-    ASSERT_NE(result, result2);
-}
-
-TEST(ParsecResultTest, Success)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    ASSERT_TRUE(result.success());
-
-    result = parsec::Result<resT> {valTest, {false, 0, {}, {}}};
-    ASSERT_FALSE(result.success());
-}
-
-TEST(ParsecResultTest, Failure)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    ASSERT_FALSE(result.failure());
-
-    result = parsec::Result<resT> {valTest, {false, 0, {}, {}}};
-    ASSERT_TRUE(result.failure());
-}
-
-TEST(ParsecResultTest, ValueRef)
-{
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    ASSERT_EQ(valTest, result.value());
-}
-
-TEST(ParsecResultTest, ValueMove)
-{
-    auto getVal = [](resT&& val)
-    {
-        return val;
-    };
-    parsec::Result<resT> result {valTest, {true, 0, {}, {}}};
-    ASSERT_EQ(valTest, getVal(result.value()));
-}
-
-TEST(ParsecResultTest, Error)
-{
-    parsec::Result<resT> result {{}, {false, 0, "message", {}}};
-    ASSERT_EQ("message", result.error());
-}
-
-TEST(ParsecResultTest, Index)
-{
-    parsec::Result<resT> result {{}, {false, 0, "message", {}}};
-    ASSERT_EQ(0, result.index());
-}
-
-TEST(ParsecResultTest, MakeSuccess)
-{
-    parsec::Result<resT> result;
-    parsec::Trace trace {true, 0, {}, {}};
-    ASSERT_NO_THROW(result = parsec::makeSuccess(int {valTest}, 0, {}, {}));
-    ASSERT_EQ(valTest, result.value());
-    ASSERT_EQ(trace, result.trace());
-}
-
-TEST(ParsecResultTest, MakeError)
-{
-    parsec::Result<resT> result;
-    parsec::Trace trace {false, 0, "message", {}};
-    ASSERT_NO_THROW(result = parsec::makeError<resT>("message", 0, {}));
-    ASSERT_EQ(trace, result.trace());
-}
-
-/****************************************************************************************/
-// Parser combinator tests
-/****************************************************************************************/
 TEST(ParsecCombinatorTest, Optional)
 {
     parsec::Result<resT> result;
@@ -296,7 +46,7 @@ TEST(ParsecCombinatorTest, Optional)
     ASSERT_NO_THROW(result = optP("test", 0));
     ASSERT_TRUE(result.success());
     ASSERT_EQ(pResult.value(), result.value());
-    ASSERT_EQ(expectedTrace, result.trace().innerTraces().value()[0]);
+    ASSERT_EQ(expectedTrace, *result.trace().innerTraces().value().front());
     ASSERT_EQ(expectedTrace.index(), result.index());
 
     p = getErrorParser();
@@ -304,7 +54,7 @@ TEST(ParsecCombinatorTest, Optional)
     expectedTrace = p("test", 0).trace();
     ASSERT_NO_THROW(result = optP("text", 0));
     ASSERT_TRUE(result.success());
-    ASSERT_EQ(expectedTrace, result.trace().innerTraces().value()[0]);
+    ASSERT_EQ(expectedTrace, *result.trace().innerTraces().value().front());
     ASSERT_EQ(0, result.index());
 }
 
@@ -320,8 +70,10 @@ TEST(ParsecCombinatorTest, OperatorLeftShift)
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_TRUE(result.success());
     ASSERT_EQ(result.value(), resultL.value());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    auto it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    it++;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(expectedTraceR.index(), result.index());
 
     pL = getErrorParser();
@@ -331,7 +83,8 @@ TEST(ParsecCombinatorTest, OperatorLeftShift)
     p = pL << pR;
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_FALSE(result.success());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
     ASSERT_EQ(expectedTraceL.index(), result.index());
 
     pL = getSuccessParser();
@@ -341,8 +94,10 @@ TEST(ParsecCombinatorTest, OperatorLeftShift)
     p = pL << pR;
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_FALSE(result.success());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    it++;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(expectedTraceR.index(), result.index());
 }
 
@@ -359,8 +114,10 @@ TEST(ParsecCombinatorTest, OperatorRightShift)
     ASSERT_TRUE(result.success());
     ASSERT_EQ(result.value(), resultR.value());
     ASSERT_EQ(2, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    auto it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    ++it;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(expectedTraceR.index(), result.index());
 
     pL = getErrorParser();
@@ -371,7 +128,8 @@ TEST(ParsecCombinatorTest, OperatorRightShift)
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_FALSE(result.success());
     ASSERT_EQ(1, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
     ASSERT_EQ(expectedTraceL.index(), result.index());
 
     pL = getSuccessParser();
@@ -382,8 +140,10 @@ TEST(ParsecCombinatorTest, OperatorRightShift)
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_FALSE(result.success());
     ASSERT_EQ(2, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    ++it;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(expectedTraceR.index(), result.index());
 }
 
@@ -403,7 +163,8 @@ TEST(ParsecCombinatorTest, OperatorOr)
     ASSERT_TRUE(result.success());
     ASSERT_EQ(result.value(), resultL.value());
     ASSERT_EQ(1, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
+    auto it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
     ASSERT_EQ(expectedTraceL.index(), result.index());
 
     // bad | ok
@@ -415,8 +176,10 @@ TEST(ParsecCombinatorTest, OperatorOr)
     ASSERT_TRUE(result.success());
     ASSERT_EQ(result.value(), resultR.value());
     ASSERT_EQ(2, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    ++it;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(expectedTraceR.index(), result.index());
 
     // bad | bad
@@ -427,8 +190,10 @@ TEST(ParsecCombinatorTest, OperatorOr)
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_FALSE(result.success());
     ASSERT_EQ(2, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    ++it;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(0, result.index());
 }
 
@@ -449,8 +214,10 @@ TEST(ParsecCombinatorTest, OperatorAnd)
     ASSERT_EQ(std::get<0>(result.value()), resultL.value());
     ASSERT_EQ(std::get<1>(result.value()), resultR.value());
     ASSERT_EQ(2, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    auto it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    ++it;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(expectedTraceR.index(), result.index());
 
     // bad & ok
@@ -461,7 +228,8 @@ TEST(ParsecCombinatorTest, OperatorAnd)
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_FALSE(result.success());
     ASSERT_EQ(1, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
     ASSERT_EQ(expectedTraceL.index(), result.index());
 
     // ok & bad
@@ -475,8 +243,10 @@ TEST(ParsecCombinatorTest, OperatorAnd)
     ASSERT_NO_THROW(result = p("test", 0));
     ASSERT_FALSE(result.success());
     ASSERT_EQ(2, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTraceL, result.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTraceR, result.trace().innerTraces().value()[1]);
+    it = result.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTraceL, **it);
+    ++it;
+    ASSERT_EQ(expectedTraceR, **it);
     ASSERT_EQ(expectedTraceR.index(), result.index());
 }
 
@@ -497,7 +267,8 @@ TEST(ParsecCombinatorTest, Fmap)
     ASSERT_TRUE(resultFmap.success());
     ASSERT_EQ(f(result.value()), resultFmap.value());
     ASSERT_EQ(1, resultFmap.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTrace, resultFmap.trace().innerTraces().value()[0]);
+    auto it = resultFmap.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTrace, **it);
     ASSERT_EQ(expectedTrace.index(), resultFmap.index());
 
     // bad
@@ -508,7 +279,8 @@ TEST(ParsecCombinatorTest, Fmap)
     ASSERT_NO_THROW(resultFmap = pFmap("test", 0));
     ASSERT_FALSE(resultFmap.success());
     ASSERT_EQ(1, resultFmap.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTrace, resultFmap.trace().innerTraces().value()[0]);
+    it = resultFmap.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTrace, **it);
     ASSERT_EQ(expectedTrace.index(), resultFmap.index());
 }
 
@@ -530,8 +302,10 @@ TEST(ParsecCombinatorTest, MonadicBinding)
     ASSERT_TRUE(resultBind.success());
     ASSERT_EQ(resultBind.value(), f(result.value())("test", result.index()).value());
     ASSERT_EQ(2, resultBind.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTrace, resultBind.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTrace2, resultBind.trace().innerTraces().value()[1]);
+    auto it = resultBind.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTrace, **it);
+    ++it;
+    ASSERT_EQ(expectedTrace2, **it);
     ASSERT_EQ(expectedTrace2.index(), resultBind.index());
 
     // bad p
@@ -542,7 +316,8 @@ TEST(ParsecCombinatorTest, MonadicBinding)
     ASSERT_NO_THROW(resultBind = pBind("test", 0));
     ASSERT_FALSE(resultBind.success());
     ASSERT_EQ(1, resultBind.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTrace, resultBind.trace().innerTraces().value()[0]);
+    it = resultBind.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTrace, **it);
     ASSERT_EQ(expectedTrace.index(), resultBind.index());
 
     // bad f
@@ -558,8 +333,10 @@ TEST(ParsecCombinatorTest, MonadicBinding)
     ASSERT_NO_THROW(resultBind = pBind("test", 0));
     ASSERT_FALSE(resultBind.success());
     ASSERT_EQ(2, resultBind.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTrace, resultBind.trace().innerTraces().value()[0]);
-    ASSERT_EQ(expectedTrace2, resultBind.trace().innerTraces().value()[1]);
+    it = resultBind.trace().innerTraces().value().begin();
+    ASSERT_EQ(expectedTrace, **it);
+    ++it;
+    ASSERT_EQ(expectedTrace2, **it);
     ASSERT_EQ(expectedTrace2.index(), resultBind.index());
 }
 
@@ -574,16 +351,16 @@ TEST(ParsecCombinatorTest, Many)
     ASSERT_TRUE(result.success());
     ASSERT_EQ(5, result.value().size());
     ASSERT_EQ(6, result.trace().innerTraces().value().size());
+    auto it = result.trace().innerTraces().value().begin();
     for (auto i = 0; i < 5; ++i)
     {
         ASSERT_EQ(i, result.value().front());
         result.value().pop_front();
-
-        ASSERT_EQ(parsec::Trace(true, i + 1, {}, {}),
-                  result.trace().innerTraces().value()[i]);
+        ASSERT_EQ(parsec::Trace(true, i + 1, {}, {}), **it);
+        ++it;
     }
     ASSERT_EQ(parsec::Trace(false, 5, "error", {}),
-              result.trace().innerTraces().value()[5]);
+              *result.trace().innerTraces().value().back());
 
     // many first error
     p = getErrorParser();
@@ -594,7 +371,7 @@ TEST(ParsecCombinatorTest, Many)
     ASSERT_TRUE(result.success());
     ASSERT_EQ(0, result.value().size());
     ASSERT_EQ(1, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTrace, result.trace().innerTraces().value()[0]);
+    ASSERT_EQ(expectedTrace, *result.trace().innerTraces().value().front());
     ASSERT_EQ(expectedTrace.index(), result.index());
 }
 
@@ -609,16 +386,17 @@ TEST(ParsecCombinatorTest, Many1)
     ASSERT_TRUE(result.success());
     ASSERT_EQ(5, result.value().size());
     ASSERT_EQ(6, result.trace().innerTraces().value().size());
+    auto it = result.trace().innerTraces().value().begin();
     for (auto i = 0; i < 5; ++i)
     {
         ASSERT_EQ(i, result.value().front());
         result.value().pop_front();
 
-        ASSERT_EQ(parsec::Trace(true, i + 1, {}, {}),
-                  result.trace().innerTraces().value()[i]);
+        ASSERT_EQ(parsec::Trace(true, i + 1, {}, {}), **it);
+        ++it;
     }
     ASSERT_EQ(parsec::Trace(false, 5, "error", {}),
-              result.trace().innerTraces().value()[5]);
+              *result.trace().innerTraces().value().back());
 
     // many1 first error
     p = getErrorParser();
@@ -628,7 +406,7 @@ TEST(ParsecCombinatorTest, Many1)
     ASSERT_NO_THROW(result = many1P("01234", 0));
     ASSERT_FALSE(result.success());
     ASSERT_EQ(1, result.trace().innerTraces().value().size());
-    ASSERT_EQ(expectedTrace, result.trace().innerTraces().value()[0]);
+    ASSERT_EQ(expectedTrace, *result.trace().innerTraces().value().front());
     ASSERT_EQ(expectedTrace.index(), result.index());
 }
